修改创建WenetSpeech方式

+++ tools/create_wenetspeech_data.py
@@ -5,11 +5,13 @@ import threading
 
 import ijson
 from pydub import AudioSegment
+from tqdm import tqdm
 
 parser = argparse.ArgumentParser(description=__doc__)
-parser.add_argument('--wenetspeech_json',  type=str,    default='/media/wenetspeech/WenetSpeech.json',  help="WenetSpeech的标注json文件路径")
+parser.add_argument('--wenetspeech_json',  type=str,    default='F:\音频数据\WenetSpeech数据集/WenetSpeech.json',  help="WenetSpeech的标注json文件路径")
 parser.add_argument('--annotation_dir',    type=str,    default='../dataset/annotation/',    help="存放数量列表的文件夹路径")
-parser.add_argument('--num_workers',       type=int,    default=8,                           help="处理数据的线程数量")
+parser.add_argument('--to_wav',            type=bool,   default=False,                       help="是否把opus格式转换为wav格式，以空间换时间")
+parser.add_argument('--num_workers',       type=int,    default=8,                           help="把opus格式转换为wav格式的线程数量")
 args = parser.parse_args()
 
 
@@ -27,7 +29,7 @@ threadLock = threading.Lock()
 threads = []
 
 
-class myThread (threading.Thread):
+class myThread(threading.Thread):
     def __init__(self, threadID, data):
         threading.Thread.__init__(self)
         self.threadID = threadID
@@ -117,24 +119,48 @@ def get_data(wenetspeech_json):
 def main():
     all_data = get_data(args.wenetspeech_json)
     print(f'总数据量为：{len(all_data)}')
-    
-    chunk_len = len(all_data) // args.num_workers
-    for i in range(args.num_workers):
-        sub_data = all_data[i * chunk_len: (i + 1) * chunk_len]
-        thread = myThread(i, sub_data)
-        thread.start()
-        threads.append(thread)
+    if args.to_wav:
+        text = input(f'音频文件将会转换为wav格式，这个过程可能很长，而且最终文件大小接近5.5T，是否继续？(y/n)')
+        if text is None or text != 'y':
+            return
+        chunk_len = len(all_data) // args.num_workers
+        for i in range(args.num_workers):
+            sub_data = all_data[i * chunk_len: (i + 1) * chunk_len]
+            thread = myThread(i, sub_data)
+            thread.start()
+            threads.append(thread)
 
-    # 等待所有线程完成
-    for t in threads:
-        t.join()
+        # 等待所有线程完成
+        for t in threads:
+            t.join()
 
+        # 复制标注文件，因为有些标注文件已经转换为wav文件
+        input_dir = os.path.dirname(args.wenetspeech_json)
+        shutil.copy(train_list_path, os.path.join(input_dir, 'wenetspeech_train.json'))
+        shutil.copy(test_list_path, os.path.join(input_dir, 'wenetspeech_test.json'))
+    else:
+        text = input(f'将直接使用opus，值得注意的是opus读取速度会比wav格式慢很多，是否继续？(y/n)')
+        if text is None or text != 'y':
+            return
+        for data in tqdm(all_data):
+            long_audio_path, segments_lists = data
+            for segment_file in segments_lists:
+                start_time = float(segment_file['begin_time'])
+                end_time = float(segment_file['end_time'])
+                text = segment_file['text']
+                confidence = segment_file['confidence']
+                if confidence < 0.95: continue
+                line = dict(audio_filepath=long_audio_path,
+                            text=text,
+                            duration=round(end_time - start_time, 3),
+                            start_time=round(start_time, 3),
+                            end_time=round(end_time, 3))
+                if long_audio_path.split('/')[-4] != 'train':
+                    f_ann_test.write('{}\n'.format(str(line).replace("'", '"')))
+                else:
+                    f_ann.write('{}\n'.format(str(line).replace("'", '"')))
     f_ann.close()
     f_ann_test.close()
-    # 复制标注文件
-    input_dir = os.path.dirname(args.wenetspeech_json)
-    shutil.copy(train_list_path, os.path.join(input_dir, 'wenetspeech_train.json'))
-    shutil.copy(test_list_path, os.path.join(input_dir, 'wenetspeech_test.json'))
 
 
 if __name__ == '__main__':
